2025 11 3
검사 바로 고

+++ src/Monster.java
@@ -4,7 +4,7 @@ public class Monster {
     String name;
     String type;
     int attack;
-    int hp;
+    double hp;
     int level;
     int rewardExp;
     Random random = new Random();
